Some compiler warnings removed.
git-svn-id: https://svn.aros.org/svn/aros/trunk@33418 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ workbench/tools/debug/sashimi/sashimi.c
@@ -143,9 +143,9 @@ STATIC struct SashimiResource * GlobalSashimiResource;
 
 /****************************************************************************/
 
-#define LVORawIOInit	 (-84 * LIB_VECTSIZE)
-#define LVORawMayGetChar (-85 * LIB_VECTSIZE)
-#define LVORawPutChar	 (-86 * LIB_VECTSIZE)
+#define LVORawIOInit	 (-84 * (WORD)LIB_VECTSIZE)
+#define LVORawMayGetChar (-85 * (WORD)LIB_VECTSIZE)
+#define LVORawPutChar	 (-86 * (WORD)LIB_VECTSIZE)
 
 /****************************************************************************/
